requirements.txt
Django==3.2.9
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-filter==21.1
django-js-asset==1.2.2
Pillow==8.4.0
python-slugify==5.0.2
text-unidecode==1.3
